ps_bq_bridge/requirements.txt
astroid==2.4.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-bigquery==2.6.1
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.0.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libcst==0.3.15
mccabe==0.6.1
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wrapt==1.12.1
